requirements.txt
Pillow
PyMuPDF
pdfplumber
python-docx
openpyxl
pypdf
reportlab
python-pptx
xlrd
pytesseract
